fix: cast input dtype in PatchEmbed and FluxTransformer to prevent dtype mismatch
When running SD3/FLUX models on CPU with torch_dtype=float16, the
projection layers (Conv2d in PatchEmbed, Linear in FluxTransformer2DModel)
may have float32 weights/bias while receiving half-precision inputs,
causing 'Input type (c10::Half) and bias type (float) should be the same'.

Cast inputs to match layer weight dtype before calling the projection.

Fixes #13300

+++ src/diffusers/models/embeddings.py
@@ -556,7 +556,7 @@ class PatchEmbed(nn.Module):
             height, width = latent.shape[-2:]
         else:
             height, width = latent.shape[-2] // self.patch_size, latent.shape[-1] // self.patch_size
-        latent = self.proj(latent)
+        latent = self.proj(latent.to(self.proj.weight.dtype))
         if self.flatten:
             latent = latent.flatten(2).transpose(1, 2)  # BCHW -> BNC
         if self.layer_norm:
